+++ CHANGELOG.md
@@ -10,6 +10,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added
 
+- Adds HTTP support for OTLP metrics exporter. (#2022)
+
 ### Changed
 
 ### Deprecated